Allow runtime toggling of repaint debugging
I have removed both the JUCE_ENABLE_REPAINT_DEBUGGING and
JUCE_REPAINT_DEBUGGING_IS_ACTIVE macros and replaced them with a
callback that allows the user to choose whether repaint debugging is
enabled on a per-component basis.

Rationale for this is that toggling JUCE_ENABLE_REPAINT_DEBUGGING
requires a full rebuild of both user code and all JUCE modules. Often
times you just want to quickly toggle repaint debugging to check that
everything is in order. Having to totally recompile your project to do
this is both annoying and unnecessary.

This does not change the ABI of juce::Component since a static callback
is used.

There's probably a cleaner way to do this. Maybe it's better to just
have a global variable that affects all components?