Optionally add AAX RequiresChunkCallsOnMainThread
Be default, Pro Tools can call setStateInformation and
getStateInformation from arbitrary threads. The trouble with this is
that it breaks some assumptions about juce::ValueTrees being safe to
read/write to on the message thread. In particular, this complicates the
interaction between APVTS and the rest of the plugin.

Since it's a big headache to rewrite old code that assumes the value
tree is only updated on the main thread, I've added the option to enable
AAX_eProperty_RequiresChunkCallsOnMainThread. This tells compatible
versions of Pro Tools that the plugin requires setStateInformation and
getStateInformation to only be run on the main thread.

You can enable this property by setting the compile definition
JucePlugin_AAXRequiresChunkCallsOnMainThread.